requirements.txt
# METRO 3D Asset Registry API - Dependencies
# ==========================================

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0

# Authentication
python-jose[cryptography]>=3.3.0
httpx>=0.26.0

# File Handling
python-multipart>=0.0.6
aiofiles>=23.2.1

# Storage Backends
boto3>=1.34.0
azure-storage-blob>=12.19.0

# JSON-LD Support
pyld>=2.0.3
rdflib>=7.0.0

# Utilities
python-dotenv>=1.0.0

# 3D Model Parsing & Metadata Extraction
trimesh>=4.0.0
pygltflib>=1.16.0
numpy>=1.26.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
aiosqlite>=0.19.0
